moved browser specific logic from core component to browser component
